ddhcp: pracitcal block_size and network size.
For testing we set the default values to the network 10.116.112.0/18 and
the block size to 32 which results in 512 blocks. I suggest this
as a good value for freifunk kiel. For bigger networks block size
of 8 should be a good deal. We should spend some thoughts on packet
amount send during a timeout update cycle and give advice on good block
amounts.